refactor(variable): update congestion fee structures
