+++ src/solver/variable/include/antares/solver/variable/economy/hydrostorage.h
@@ -1,59 +1,245 @@
 // Copyright 2007-2026, RTE (https://www.rte-france.com)
 // SPDX-License-Identifier: MPL-2.0
 
-#pragma once
+#ifndef __SOLVER_VARIABLE_ECONOMY_HydroStorage_H__
+#define __SOLVER_VARIABLE_ECONOMY_HydroStorage_H__
 
-#include "economy_base.h"
+#include "antares/solver/variable/variable.h"
 
 namespace Antares::Solver::Variable::Economy
 {
-struct HydroStorageTraits
+struct VCardHydroStorage
 {
+    //! Caption
     static std::string Caption()
     {
         return "H. STOR";
     }
 
+    //! Unit
     static std::string Unit()
     {
         return "MWh";
     }
 
+    //! The short description of the variable
     static std::string Description()
     {
         return "Hydro Storage Generation";
     }
 
-    using ResultsType = Results<R::AllYears::Average<R::AllYears::StdDeviation<R::AllYears::Min<
-      R::AllYears::Max<>>>>>;
+    //! The expecte results
+    typedef Results<R::AllYears::Average< // The average values throughout all years
+      R::AllYears::StdDeviation<          // The standard deviation values throughout all years
+        R::AllYears::Min<                 // The minimum values throughout all years
+          R::AllYears::Max<               // The maximum values throughout all years
+            >>>>>
+      ResultsType;
 
+    //! The VCard to look for for calculating spatial aggregates
+    typedef VCardHydroStorage VCardForSpatialAggregate;
+
+    //! Data Level
+    static constexpr uint8_t categoryDataLevel = Category::DataLevel::area;
+    //! File level (provided by the type of the results)
+    static constexpr uint8_t categoryFileLevel = ResultsType::categoryFile
+                                                 & (Category::FileLevel::id
+                                                    | Category::FileLevel::va);
+    //! Precision (views)
+    static constexpr uint8_t precision = Category::all;
+    //! Indentation (GUI)
+    static constexpr uint8_t nodeDepthForGUI = +0;
+    //! Decimal precision
     static constexpr uint8_t decimal = 0;
+    //! Number of columns used by the variable (One ResultsType per column)
+    static constexpr int columnCount = 1;
+    //! The Spatial aggregation
     static constexpr uint8_t spatialAggregate = Category::spatialAggregateSum;
+    static constexpr uint8_t spatialAggregateMode = Category::spatialAggregateEachYear;
+    static constexpr uint8_t spatialAggregatePostProcessing = 0;
+    //! Intermediate values
+    static constexpr uint8_t hasIntermediateValues = 1;
+    //! Can this variable be non applicable (0 : no, 1 : yes)
+    static constexpr uint8_t isPossiblyNonApplicable = 0;
 
-    static bool checkCondition(const State&)
-    {
-        return true;
-    }
+    typedef IntermediateValues IntermediateValuesBaseType;
+    typedef std::vector<IntermediateValues> IntermediateValuesType;
 
-    static double value(const State& state)
-    {
-        return state.hourlyResults->TurbinageHoraire[state.hourInTheWeek];
-    }
+    using IntermediateValuesTypeForSpatialAg = std::unique_ptr<IntermediateValuesBaseType[]>;
 
-    static void computeStats(IntermediateValues& intermediateValues)
-    {
-        intermediateValues.computeStatisticsForTheCurrentYear();
-    }
-};
-
-using VCardHydroStorage = VCard_Base<HydroStorageTraits>;
+}; // class VCard
 
 /*!
 ** \brief C02 Average value of the overrall HydroStorage emissions expected from all
 **   the thermal dispatchable clusters
 */
 template<class NextT = Container::EndOfList>
-using HydroStorage = Economy_Base<HydroStorageTraits, NextT>;
+class HydroStorage: public Variable::IVariable<HydroStorage<NextT>, NextT, VCardHydroStorage>
+{
+public:
+    //! Type of the next static variable
+    typedef NextT NextType;
+    //! VCard
+    typedef VCardHydroStorage VCardType;
+    //! Ancestor
+    typedef Variable::IVariable<HydroStorage<NextT>, NextT, VCardType> AncestorType;
+
+    //! List of expected results
+    typedef typename VCardType::ResultsType ResultsType;
+
+    typedef VariableAccessor<ResultsType, VCardType::columnCount> VariableAccessorType;
+
+    enum
+    {
+        //! How many items have we got
+        count = 1 + NextT::count,
+    };
+
+    template<int CDataLevel, int CFile>
+    struct Statistics
+    {
+        enum
+        {
+            count = ((VCardType::categoryDataLevel & CDataLevel
+                      && VCardType::categoryFileLevel & CFile)
+                       ? (NextType::template Statistics<CDataLevel, CFile>::count
+                          + VCardType::columnCount * ResultsType::count)
+                       : NextType::template Statistics<CDataLevel, CFile>::count),
+        };
+    };
+
+public:
+    void initializeFromStudy(Data::Study& study)
+    {
+        pNbYearsParallel = study.maxNbYearsInParallel;
+
+        InitializeResultsFromStudy(AncestorType::pResults, study);
+
+        pValuesForTheCurrentYear.resize(pNbYearsParallel);
+        for (unsigned int numSpace = 0; numSpace < pNbYearsParallel; numSpace++)
+        {
+            pValuesForTheCurrentYear[numSpace].initializeFromStudy(study);
+        }
+
+        // Next
+        NextType::initializeFromStudy(study);
+    }
+
+    template<class R>
+    static void InitializeResultsFromStudy(R& results, Data::Study& study)
+    {
+        VariableAccessorType::InitializeAndReset(results, study);
+    }
+
+    void initializeFromArea(Data::Study* study, Data::Area* area)
+    {
+        // Next
+        NextType::initializeFromArea(study, area);
+    }
+
+    void initializeFromLink(Data::Study* study, Data::AreaLink* link)
+    {
+        // Next
+        NextType::initializeFromAreaLink(study, link);
+    }
+
+    void simulationBegin()
+    {
+        for (unsigned int numSpace = 0; numSpace < pNbYearsParallel; numSpace++)
+        {
+            pValuesForTheCurrentYear[numSpace].reset();
+        }
+        // Next
+        NextType::simulationBegin();
+    }
+
+    void simulationEnd()
+    {
+        NextType::simulationEnd();
+    }
+
+    void yearBegin(unsigned int year, unsigned int numSpace)
+    {
+        // Reset the values for the current year
+        pValuesForTheCurrentYear[numSpace].reset();
+
+        // Next variable
+        NextType::yearBegin(year, numSpace);
+    }
+
+    void yearEndBuild(State& state, unsigned int year, unsigned int numSpace)
+    {
+        // Next variable
+        NextType::yearEndBuild(state, year, numSpace);
+    }
+
+    void yearEnd(unsigned int year, unsigned int numSpace)
+    {
+        // Compute all statistics for the current year (daily, weekly, monthly)
+        pValuesForTheCurrentYear[numSpace].computeStatisticsForTheCurrentYear();
+
+        // Next variable
+        NextType::yearEnd(year, numSpace);
+    }
+
+    void computeSummary(unsigned int year, unsigned int numSpace)
+    {
+        // Merge all those values with the global results
+        AncestorType::pResults.merge(year, pValuesForTheCurrentYear[numSpace]);
+
+        // Next variable
+        NextType::computeSummary(year, numSpace);
+    }
+
+    void hourBegin(unsigned int hourInTheYear)
+    {
+        // Next variable
+        NextType::hourBegin(hourInTheYear);
+    }
+
+    void hourForEachArea(State& state, unsigned int numSpace)
+    {
+        // Total HydroStorage emissions
+        pValuesForTheCurrentYear[numSpace].hour[state.hourInTheYear] =
+          // Current Hydro Storage generation
+          state.hourlyResults->TurbinageHoraire[state.hourInTheWeek];
+
+        // Next variable
+        NextType::hourForEachArea(state, numSpace);
+    }
+
+    Antares::Memory::Stored<double>::ConstReturnType retrieveRawHourlyValuesForCurrentYear(
+      unsigned int,
+      unsigned int numSpace) const
+    {
+        return pValuesForTheCurrentYear[numSpace].hour;
+    }
+
+    void localBuildAnnualSurveyReport(SurveyResults& results,
+                                      int fileLevel,
+                                      int precision,
+                                      unsigned int numSpace) const
+    {
+        // Initializing external pointer on current variable non applicable status
+        results.isCurrentVarNA = AncestorType::isNonApplicable;
+
+        if (AncestorType::isPrinted[0])
+        {
+            // Write the data for the current year
+            results.variableCaption = VCardType::Caption();
+            results.variableUnit = VCardType::Unit();
+            pValuesForTheCurrentYear[numSpace]
+              .template buildAnnualSurveyReport<VCardType>(results, fileLevel, precision);
+        }
+    }
+
+private:
+    //! Intermediate values for each year
+    typename VCardType::IntermediateValuesType pValuesForTheCurrentYear;
+    unsigned int pNbYearsParallel;
+
+}; // class HydroStorage
 
 } // namespace Antares::Solver::Variable::Economy
 
+#endif // __SOLVER_VARIABLE_ECONOMY_HydroStorage_H__
