+++ src/solver/variable/include/antares/solver/variable/economy/links/flowLinear.h
@@ -4,31 +4,27 @@
 #ifndef __SOLVER_VARIABLE_ECONOMY_FlowLinear_H__
 #define __SOLVER_VARIABLE_ECONOMY_FlowLinear_H__
 
-#include "../../variable.h"
+#include "EconomyLinkBase.h"
 
 namespace Antares::Solver::Variable::Economy
 {
-struct VCardFlowLinear
+struct FlowLinearTraits
 {
-    //! Caption
     static std::string Caption()
     {
         return "FLOW LIN.";
     }
 
-    //! Unit
     static std::string Unit()
     {
         return "MWh";
     }
 
-    //! The short description of the variable
     static std::string Description()
     {
         return "Flow assessed, over all MC years, through linear optimization";
     }
 
-    //! The expecte results
     typedef Results<R::AllYears::Average< // The average values throughout all years
       R::AllYears::StdDeviation<          // The standard deviation values throughout all years
         R::AllYears::Min<                 // The minimum values throughout all years
@@ -36,173 +32,16 @@ struct VCardFlowLinear
             >>>>>
       ResultsType;
 
-    //! Data Level
-    static constexpr uint8_t categoryDataLevel = Category::DataLevel::link;
-    //! File level (provided by the type of the results)
-    static constexpr uint8_t categoryFileLevel = ResultsType::categoryFile
-                                                 & (Category::FileLevel::id
-                                                    | Category::FileLevel::va);
-    //! Precision (views)
-    static constexpr uint8_t precision = Category::all;
-    //! Indentation (GUI)
-    static constexpr uint8_t nodeDepthForGUI = +0;
-    //! Decimal precision
-    static constexpr uint8_t decimal = 0;
-    //! Number of columns used by the variable (One ResultsType per column)
-    static constexpr int columnCount = 1;
-    //! The Spatial aggregation
-    static constexpr uint8_t spatialAggregate = Category::spatialAggregateSum;
-    static constexpr uint8_t spatialAggregateMode = Category::spatialAggregateEachYear;
-    static constexpr uint8_t spatialAggregatePostProcessing = 0;
-    //! Intermediate values
-    static constexpr uint8_t hasIntermediateValues = 1;
-    //! Can this variable be non applicable (0 : no, 1 : yes)
-    static constexpr uint8_t isPossiblyNonApplicable = 0;
-
-    typedef IntermediateValues IntermediateValuesBaseType;
-    typedef std::vector<IntermediateValues> IntermediateValuesType;
-
-}; // class VCard
-
-/*!
-** \brief Marginal FlowLinear
-*/
-template<class NextT = Container::EndOfList>
-class FlowLinear: public Variable::IVariable<FlowLinear<NextT>, NextT, VCardFlowLinear>
-{
-public:
-    //! Type of the next static variable
-    typedef NextT NextType;
-    //! VCard
-    typedef VCardFlowLinear VCardType;
-    //! Ancestor
-    typedef Variable::IVariable<FlowLinear<NextT>, NextT, VCardType> AncestorType;
-
-    //! List of expected results
-    typedef typename VCardType::ResultsType ResultsType;
-
-    typedef VariableAccessor<ResultsType, VCardType::columnCount> VariableAccessorType;
-
-    enum
+    static double hourValue(State& state)
     {
-        //! How many items have we got
-        count = 1 + NextT::count,
-    };
-
-    template<int CDataLevel, int CFile>
-    struct Statistics
-    {
-        enum
-        {
-            count = ((VCardType::categoryDataLevel & CDataLevel
-                      && VCardType::categoryFileLevel & CFile)
-                       ? (NextType::template Statistics<CDataLevel, CFile>::count
-                          + VCardType::columnCount * ResultsType::count)
-                       : NextType::template Statistics<CDataLevel, CFile>::count),
-        };
-    };
-
-public:
-    void initializeFromStudy(Data::Study& study)
-    {
-        pNbYearsParallel = study.maxNbYearsInParallel;
-
-        // Average on all years
-        AncestorType::pResults.initializeFromStudy(study);
-        AncestorType::pResults.reset();
-
-        // Intermediate values
-        pValuesForTheCurrentYear.resize(pNbYearsParallel);
-        for (unsigned int numSpace = 0; numSpace < pNbYearsParallel; numSpace++)
-        {
-            pValuesForTheCurrentYear[numSpace].initializeFromStudy(study);
-        }
-
-        // Next
-        NextType::initializeFromStudy(study);
+        return state.ntc.ValeurDuFlux[state.link->index];
     }
 
-    void initializeFromArea(Data::Study* study, Data::Area* area)
-    {
-        // Next
-        NextType::initializeFromArea(study, area);
-    }
-
-    void initializeFromAreaLink(Data::Study* study, Data::AreaLink* link)
-    {
-        // Next
-        NextType::initializeFromAreaLink(study, link);
-    }
-
-    void simulationBegin()
-    {
-        for (unsigned int numSpace = 0; numSpace < pNbYearsParallel; numSpace++)
-        {
-            pValuesForTheCurrentYear[numSpace].reset();
-        }
-        // Next
-        NextType::simulationBegin();
-    }
-
-    void simulationEnd()
-    {
-        NextType::simulationEnd();
-    }
-
-    void yearBegin(uint year, unsigned int numSpace)
-    {
-        // Reset
-        pValuesForTheCurrentYear[numSpace].reset();
-        // Next variable
-        NextType::yearBegin(year, numSpace);
-    }
-
-    void yearEndBuild(State& state, unsigned int year, unsigned int numSpace)
-    {
-        // Next variable
-        NextType::yearEndBuild(state, year, numSpace);
-    }
-
-    void yearEnd(unsigned int year, unsigned int numSpace)
-    {
-        // Compute all statistics for the current year (daily,weekly,monthly)
-        pValuesForTheCurrentYear[numSpace].computeStatisticsForTheCurrentYear();
-
-        // Next variable
-        NextType::yearEnd(year, numSpace);
-    }
-
-    void computeSummary(unsigned int year, unsigned int numSpace)
-    {
-        // Merge all those values with the global results
-        AncestorType::pResults.merge(year, pValuesForTheCurrentYear[numSpace]);
-
-        // Next variable
-        NextType::computeSummary(year, numSpace);
-    }
-
-    void hourBegin(uint hourInTheYear)
-    {
-        // Next variable
-        NextType::hourBegin(hourInTheYear);
-    }
-
-    void hourForEachArea(State& state, unsigned int numSpace)
-    {
-        // Next variable
-        NextType::hourForEachArea(state, numSpace);
-    }
-
-    void hourForEachLink(State& state, unsigned int numSpace)
-    {
-        // Flow assessed over all MC years (linear)
-        pValuesForTheCurrentYear[numSpace].hour[state.hourInTheYear] += state.ntc.ValeurDuFlux
-                                                                          [state.link->index];
-        // Next item in the list
-        NextType::hourForEachLink(state, numSpace);
-    }
-
-    void buildDigest(SurveyResults& results, int digestLevel, int dataLevel) const
+    template<class R>
+    static void buildDigest(SurveyResults& results,
+                            int digestLevel,
+                            int dataLevel,
+                            const R& pResults)
     {
         if (dataLevel & Category::DataLevel::link)
         {
@@ -210,48 +49,22 @@ public:
             {
                 results.data.matrix
                   .entry[results.data.link->from->index][results.data.link->with->index]
-                  = AncestorType::pResults.avgdata.year;
+                  = pResults.avgdata.year;
                 results.data.matrix
                   .entry[results.data.link->with->index][results.data.link->from->index]
-                  = -AncestorType::pResults.avgdata.year;
+                  = -pResults.avgdata.year;
             }
         }
-
-        // Next
-        NextType::buildDigest(results, digestLevel, dataLevel);
     }
+};
 
-    Antares::Memory::Stored<double>::ConstReturnType retrieveRawHourlyValuesForCurrentYear(
-      uint,
-      uint numSpace) const
-    {
-        return pValuesForTheCurrentYear[numSpace].hour;
-    }
+/*!
+** \brief Marginal FlowLinear
+*/
+template<class NextT = Container::EndOfList>
+using FlowLinear = EconomyLinkBase<FlowLinearTraits, NextT>;
 
-    void localBuildAnnualSurveyReport(SurveyResults& results,
-                                      int fileLevel,
-                                      int precision,
-                                      uint numSpace) const
-    {
-        // Initializing external pointer on current variable non applicable status
-        results.isCurrentVarNA = AncestorType::isNonApplicable;
-
-        if (AncestorType::isPrinted[0])
-        {
-            // Write the data for the current year
-            results.variableCaption = VCardType::Caption();
-            results.variableUnit = VCardType::Unit();
-            pValuesForTheCurrentYear[numSpace]
-              .template buildAnnualSurveyReport<VCardType>(results, fileLevel, precision);
-        }
-    }
-
-private:
-    //! Intermediate values for each year
-    typename VCardType::IntermediateValuesType pValuesForTheCurrentYear;
-    unsigned int pNbYearsParallel;
-
-}; // class FlowLinear
+using VCardFlowLinear = VCardEconomyLinkBase<FlowLinearTraits>;
 
 } // namespace Antares::Solver::Variable::Economy
