+++ src/solver/variable/include/antares/solver/variable/economy/links/marginalCost.h
@@ -6,31 +6,27 @@
 
 #include <cmath>
 
-#include "../../variable.h"
+#include "links_base.h"
 
 namespace Antares::Solver::Variable::Economy
 {
-struct VCardMarginalCost
+struct MarginalCostTraits
 {
-    //! Caption
     static std::string Caption()
     {
         return "MARG. COST";
     }
 
-    //! Unit
     static std::string Unit()
     {
         return "Euro/MW";
     }
 
-    //! The short description of the variable
     static std::string Description()
     {
         return "Decrease of the overall operating cost expected by a 1MW capacity reinforcement";
     }
 
-    //! The expecte results
     typedef Results<R::AllYears::Average< // The average values throughout all years
       R::AllYears::StdDeviation<          // The standard deviation values throughout all years
         R::AllYears::Min<                 // The minimum values throughout all years
@@ -38,220 +34,38 @@ struct VCardMarginalCost
             >>>>>
       ResultsType;
 
-    //! Data Level
-    static constexpr uint8_t categoryDataLevel = Category::DataLevel::link;
-    //! File level (provided by the type of the results)
-    static constexpr uint8_t categoryFileLevel = ResultsType::categoryFile
-                                                 & (Category::FileLevel::id
-                                                    | Category::FileLevel::va);
-    //! Precision (views)
-    static constexpr uint8_t precision = Category::all;
-    //! Indentation (GUI)
-    static constexpr uint8_t nodeDepthForGUI = +0;
-    //! Decimal precision
     static constexpr uint8_t decimal = 2;
-    //! Number of columns used by the variable (One ResultsType per column)
-    static constexpr int columnCount = 1;
-    //! The Spatial aggregation
-    static constexpr uint8_t spatialAggregate = Category::spatialAggregateSum;
-    static constexpr uint8_t spatialAggregateMode = Category::spatialAggregateEachYear;
-    static constexpr uint8_t spatialAggregatePostProcessing = 0;
-    //! Intermediate values
-    static constexpr uint8_t hasIntermediateValues = 1;
-    //! Can this variable be non applicable (0 : no, 1 : yes)
-    static constexpr uint8_t isPossiblyNonApplicable = 0;
 
-    typedef IntermediateValues IntermediateValuesBaseType;
-    typedef std::vector<IntermediateValues> IntermediateValuesType;
+    static void computeStats(IntermediateValues& iv)
+    {
+        iv.computeAveragesForCurrentYearFromHourlyResults();
+    }
 
-}; // class VCard
+    static void hourForEachLink(IntermediateValues& iv, State& state)
+    {
+        iv[state.hourInTheYear] = std::abs(
+          state.problemeHebdo->VariablesDualesDesContraintesDeNTC[state.hourInTheWeek]
+            .VariableDualeParInterconnexion[state.link->index]);
+
+        const double flow = state.problemeHebdo->ValeursDeNTC[state.hourInTheWeek]
+                              .ValeurDuFlux[state.link->index];
+        const double flow_lowerbound = -state.problemeHebdo->ValeursDeNTC[state.hourInTheWeek]
+                                          .ValeurDeNTCExtremiteVersOrigine[state.link->index];
+        const double flow_upperbound = state.problemeHebdo->ValeursDeNTC[state.hourInTheWeek]
+                                         .ValeurDeNTCOrigineVersExtremite[state.link->index];
+
+        if (flow - 0.001 > flow_lowerbound && flow + 0.001 < flow_upperbound)
+        {
+            iv[state.hourInTheYear] = 0.;
+        }
+    }
+};
 
 /*!
 ** \brief Marginal MarginalCost
 */
 template<class NextT = Container::EndOfList>
-class MarginalCost: public Variable::IVariable<MarginalCost<NextT>, NextT, VCardMarginalCost>
-{
-public:
-    //! Type of the next static variable
-    typedef NextT NextType;
-    //! VCard
-    typedef VCardMarginalCost VCardType;
-    //! Ancestor
-    typedef Variable::IVariable<MarginalCost<NextT>, NextT, VCardType> AncestorType;
-
-    //! List of expected results
-    typedef typename VCardType::ResultsType ResultsType;
-
-    typedef VariableAccessor<ResultsType, VCardType::columnCount> VariableAccessorType;
-
-    enum
-    {
-        //! How many items have we got
-        count = 1 + NextT::count,
-    };
-
-    template<int CDataLevel, int CFile>
-    struct Statistics
-    {
-        enum
-        {
-            count = ((VCardType::categoryDataLevel & CDataLevel
-                      && VCardType::categoryFileLevel & CFile)
-                       ? (NextType::template Statistics<CDataLevel, CFile>::count
-                          + VCardType::columnCount * ResultsType::count)
-                       : NextType::template Statistics<CDataLevel, CFile>::count),
-        };
-    };
-
-public:
-    void initializeFromStudy(Data::Study& study)
-    {
-        pNbYearsParallel = study.maxNbYearsInParallel;
-
-        // Average on all years
-        AncestorType::pResults.initializeFromStudy(study);
-        AncestorType::pResults.reset();
-
-        // Intermediate values
-        pValuesForTheCurrentYear.resize(pNbYearsParallel);
-        for (unsigned int numSpace = 0; numSpace < pNbYearsParallel; numSpace++)
-        {
-            pValuesForTheCurrentYear[numSpace].initializeFromStudy(study);
-        }
-
-        // Next
-        NextType::initializeFromStudy(study);
-    }
-
-    void initializeFromArea(Data::Study* study, Data::Area* area)
-    {
-        // Next
-        NextType::initializeFromArea(study, area);
-    }
-
-    void initializeFromAreaLink(Data::Study* study, Data::AreaLink* link)
-    {
-        // Next
-        NextType::initializeFromAreaLink(study, link);
-    }
-
-    void simulationBegin()
-    {
-        // Next
-        NextType::simulationBegin();
-    }
-
-    void simulationEnd()
-    {
-        NextType::simulationEnd();
-    }
-
-    void yearBegin(uint year, unsigned int numSpace)
-    {
-        // Reset
-        pValuesForTheCurrentYear[numSpace].reset();
-        // Next variable
-        NextType::yearBegin(year, numSpace);
-    }
-
-    void yearEndBuild(State& state, unsigned int year, unsigned int numSpace)
-    {
-        // Next variable
-        NextType::yearEndBuild(state, year, numSpace);
-    }
-
-    void yearEnd(unsigned int year, unsigned int numSpace)
-    {
-        // Compute all statistics for the current year (daily,weekly,monthly)
-        pValuesForTheCurrentYear[numSpace].computeAveragesForCurrentYearFromHourlyResults();
-
-        // Next variable
-        NextType::yearEnd(year, numSpace);
-    }
-
-    void computeSummary(unsigned int year, unsigned int numSpace)
-    {
-        // Merge all those values with the global results
-        AncestorType::pResults.merge(year, pValuesForTheCurrentYear[numSpace]);
-
-        // Next variable
-        NextType::computeSummary(year, numSpace);
-    }
-
-    void hourBegin(uint hourInTheYear)
-    {
-        // Next variable
-        NextType::hourBegin(hourInTheYear);
-    }
-
-    void hourForEachArea(State& state, unsigned int numSpace)
-    {
-        // Next variable
-        NextType::hourForEachArea(state, numSpace);
-    }
-
-    void hourForEachLink(State& state, unsigned int numSpace)
-    {
-        pValuesForTheCurrentYear[numSpace][state.hourInTheYear] = std::abs(
-          state.problemeHebdo->VariablesDualesDesContraintesDeNTC[state.hourInTheWeek]
-            .VariableDualeParInterconnexion[state.link->index]);
-
-        // This value should be reset to zero if  (flow_lowerbound) < flow < (flow_upperbound) (with
-        // signed values)
-        double flow = state.problemeHebdo->ValeursDeNTC[state.hourInTheWeek]
-                        .ValeurDuFlux[state.link->index];
-        double flow_lowerbound = -state.problemeHebdo->ValeursDeNTC[state.hourInTheWeek]
-                                    .ValeurDeNTCExtremiteVersOrigine[state.link->index];
-        double flow_upperbound = state.problemeHebdo->ValeursDeNTC[state.hourInTheWeek]
-                                   .ValeurDeNTCOrigineVersExtremite[state.link->index];
-
-        if (flow - 0.001 > flow_lowerbound && flow + 0.001 < flow_upperbound)
-        {
-            pValuesForTheCurrentYear[numSpace][state.hourInTheYear] = 0.;
-        }
-
-        // Next item in the list
-        NextType::hourForEachLink(state, numSpace);
-    }
-
-    void buildDigest(SurveyResults& results, int digestLevel, int dataLevel) const
-    {
-        // Next
-        NextType::buildDigest(results, digestLevel, dataLevel);
-    }
-
-    Antares::Memory::Stored<double>::ConstReturnType retrieveRawHourlyValuesForCurrentYear(
-      uint,
-      uint numSpace) const
-    {
-        return pValuesForTheCurrentYear[numSpace].hour;
-    }
-
-    void localBuildAnnualSurveyReport(SurveyResults& results,
-                                      int fileLevel,
-                                      int precision,
-                                      uint numSpace) const
-    {
-        // Initializing external pointer on current variable non applicable status
-        results.isCurrentVarNA = AncestorType::isNonApplicable;
-
-        if (AncestorType::isPrinted[0])
-        {
-            // Write the data for the current year
-            results.variableCaption = VCardType::Caption();
-            results.variableUnit = VCardType::Unit();
-            pValuesForTheCurrentYear[numSpace]
-              .template buildAnnualSurveyReport<VCardType>(results, fileLevel, precision);
-        }
-    }
-
-private:
-    //! Intermediate values for each year
-    typename VCardType::IntermediateValuesType pValuesForTheCurrentYear;
-    unsigned int pNbYearsParallel;
-
-}; // class MarginalCost
+using MarginalCost = EconomyLink_Base<MarginalCostTraits, NextT>;
 
 } // namespace Antares::Solver::Variable::Economy
